[REF] client: group tests by behavior
